lava_dispatcher: match devices by udev attributes
This patch allows to use arbitrary udev variables to match devices in
device_info and static_info entries, and makes board_id non mandatory.

board_id, usb_vendor_id, usb_product_id, fs_label are now deprecated,
but there is a lot of code that still assumes them.

Note that in the absense of board_id, or ID_SERIAL_SHORT, devices are not
garanteed to be unique, and LAVA will not raise any exception in that
case.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>
Signed-off-by: Stevan Radakovic <stevan.radakovic@linaro.org>
Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>
Signed-off-by: Tim Jaacks <tim.jaacks@garz-fricke.com>