[RFC] lava_scheduler_app: Handle exceptions from Device.is_valid(), etc.
In previous commit, Device.is_valid() and Device.load_configuration()
methods were made to raise exceptions, so any issues with device
configuration were propagated up to their "business logic" handlers,
at which place the issue they can be reported to user in appropriate
context and detail.

This commit of the series makes the changes to that effect, changing
most of the callers of those methods to expect and handle exceptions.
Where they were handled, the scope of handling was extended, to handle
any exceptions (as indeed, the set of exceptions may change with the
evolution of applications).

In rare cases, handling still consists of ignoring exception, as in
e.g. Device.get_extends() (arguably, if device configuration is invalid,
it's fair to say that it has none of Jinja "extends" directives).

In most other cases, handling is at least logging the exception via
appropriate logger. These cases won't be visible to an end user, but
at least will be visible to administartor or first-time installer.

However, in the cases where audit trail of operations is added to
database, information from the exception is included in such audit
trail to reach users and help debugging.

For some web views, exceptions aren't handled explicitly, asssuming
that a default webapp exception handler will handle it (at least log
and report a generic error to use). Such cases actually should be
re-reviewed by the core team, and possibly better handling is added.

One last note is sadly, there's no consistency as how different
exception classes handle their str() and repr() operations. For a
particular exception, either of these may miss important information.
For exmaple, jinja2's TemplateNotFound has str() of just template
filename (so, on its own, it wouldn't be clear what the issue is with
that filename), while repr() is just "TemplateNotFound()" (no filename
included), so we need to dump both exception type name and its str().
If this patchset is accepted, that needs to be done consistently
(it's currently not), perhaps by introducing a helper function like
format_exc().

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>