Work-around django-environ unnecessary warning
django-environ warns at import time if an .env file does not exist
on the same directory as the django settings file.

The warning is unnecessary, since if you did not create the file it's
because you want to provide the environment variables directly.

To silence it, we install an empty .env file to the expected place.

Closes #485

Signed-off-by: Gustavo Noronha Silva <gustavo.noronha@collabora.com>