Optimize viewing_groups filter for TestJob
Invert the algorithm

Old: AND NOT (group_id IN non_user_groups)
New: AND (group_id IN user_groups)

Previously the filter would make two separated
joins to verify group_id is not NULL and
to actually find the viewing_groups.

To find out what groups user was not in would
be done inside python with a separated query. Now it is
inside database.

Example where user with id 2
is inside the group 45.

Before:

OR (	"lava_scheduler_app_testjob"."id" IN (
 		SELECT U0."id"
 		FROM "lava_scheduler_app_testjob" U0
 		INNER JOIN  "lava_scheduler_app_testjob_viewing_groups" U1
 		ON (U0."id" = U1."testjob_id")
 		WHERE U1."group_id" IS NOT NULL)
 	AND NOT ("lava_scheduler_app_testjob"."id" IN (
 		SELECT V1."testjob_id"
 		FROM "lava_scheduler_app_testjob_viewing_groups" V1
 		WHERE V1."group_id" IN
 			(SELECT U0."id" FROM "auth_group" U0 WHERE NOT (U0."id" IN (45))))))))

After:

OR "lava_scheduler_app_testjob"."id" IN
 	(SELECT V0."id"
 	 FROM "lava_scheduler_app_testjob" V0
 	 INNER JOIN "lava_scheduler_app_testjob_viewing_groups" V1
 	 	ON (V0."id" =  V1."testjob_id")
 	 WHERE (
 	 	V1."group_id" IS NOT NULL
 	        AND V1."group_id" IN (
 	            SELECT U0."id"
 	            FROM "auth_group" U0
 	            INNER JOIN "auth_user_groups" U1
 	            	ON (U0."id" = U1."group_id")
 	            WHERE U1."user_id" = 2)))))

Tested on a generated database.