U-Boot: Allow adding base commands
Currently it is not possible to add commands to base_uboot_commands, so change template to make this possible. This can be needed by devices that need custom commands issued before fetching boot image, e.g some commands to enable ethernet.

Default is the same behavior as previously, so it shouldn't affect any existing jobs.