deploy: Fix the issue of overlay unpack if the OS value is present
OS is an important value of a LAVA job definition which will export
the lava test shell and will distinguish which OS we are using like
debian,OE or ubuntu etc. The value will invoke the posix variable
from the script deployment_data.py and its not able to replace
the lava_test_results_dir in the context session, So the overlay unpack
is not happening in the specified custom path.

This will misplace the  lava-test-runner  as well, the patch will take the
lava_test_results_dir from the job definition if exists or else will go for the posix value.

Signed-off-by: sarath P T <sarath.pt@siemens.com>