dispatcher: tftp: Support deployment without a kernel
When LAVA is used to automate bootloader testing, the tftp deployment
action may be used to supply test scripts and data so that they can be
downloaded to the DUT during a test. In this use case, Linux may not
actually be booted and so no kernel image may be needed.

Signed-off-by: Paul Barker <paul.barker.ct@bp.renesas.com>