dispatcher/actions/download: avoid automatic gzip decompression
The requests library automatically decompresses its content if the "Content-Encoding"
header is set to a compression library available to requests.

This is explicit invoked by the requests library[1], which uses urllib3's feature[2].

In my case, LAVA fails to download a Image.gz because the server set "Content-Encoding: gzip",
which I cannot control because it comes from AWS S3.

The requests library automatically decompresses the content, and when LAVA attempts to decompress it,
an exception is raised because the content is no longer supported by gzip library.

[1] https://github.com/psf/requests/blob/1ae6fc3137a11e11565ed22436aa1e77277ac98c/src/requests/models.py#L820
[2] https://urllib3.readthedocs.io/en/stable/reference/urllib3.response.html#urllib3.response.BaseHTTPResponse.stream

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>