Add secrets filtering for job output
This is somewhat similar to how Gitlab does its secret filtering.
If a secret value is accidentally leaked in the job output it will
be replaced with `[MASKED]` string to prevent the secret from
being exposed in saved logs.

Use a global variable in `lava_common/log.py` to store the secret
values so that no `YAMLLogger` escapes filtering.