Create Job.secrets attribute as a unified secrets dictionary
Instead of having to lookup the job definition to get secrets
add `Job.secrets` attribute of `dict[str, str]` type.
Currently this dictionary is initialized in `Job.__init__` from
the job definition. In the future this can be expanded to alternative
sources of secrets. For example, `systemd-creds` support can be
added.

Switch `AvhDeploy` and `CreateOverlay` to use this new attribute.
It seems like those two actions are the only users of secrets.
The HTTP requests headers are updated with the secrets on the
lava-server before the job is downloaded.